Changing the layout a bit. Adding apft data to homepage

+++ models/apftData.go
@@ -7,7 +7,7 @@ import (
 //APFTData Model representing apft data
 type APFTData struct {
 	ID              bson.ObjectId `json:"id" bson:"_id,omitempty"`
-	PercentFail     int           `json:"percentfail" bson:"percentfail"`
-	PercentPassed   int           `json:"percentpassed" bson:"percentpassed"`
-	PercentNotTaken int           `json:"percentnottaken" bson:"percentnottaken"`
+	PercentFail     float32       `json:"percentfail" bson:"percentfail"`
+	PercentPassed   float32       `json:"percentpassed" bson:"percentpassed"`
+	PercentNotTaken float32       `json:"percentnottaken" bson:"percentnottaken"`
 }
